big add: Users model, sessions, signup, login, logout
Nominally following the old railscast on rolling your own authentication.

whee!

also:

- added bootstrap via yarn, along with jquery and popper.js
- added a scss file to load up bootstrap styles
- added rspec, wrote some tests for the User model

This was a bit of a stumper: I'm used to being able to call render in the application.html.erb layout without any issues. This seems to no longer work the same in Rails6. I needed to resort to calling it as `render template: "layouts/_partial_file"` which just seems weirder than weird.